game-telemetry/requirements.txt
google-cloud-pubsub==2.13.10
google-cloud-bigquery==3.3.5
google-cloud-storage==2.5.0